requirements.txt
fastapi
uvicorn
python-dotenv
google-api-python-client
google-auth
pypdf
sentence-transformers
pinecone
groq